+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * 【core   】     增加BeanValueProvider（issue#I5FBHV@Gitee）
 * 【core   】     Convert工具类中，新增中文大写数字金额转换为数字工具方法（pr#674@Gitee）
 * 【core   】     新增CollectorUtil.reduceListMap()（pr#676@Gitee）
+* 【core   】     新增CollectorUtil.reduceListMap()（pr#676@Gitee）
 * 
 ### 🐞Bug修复