+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 
 -------------------------------------------------------------------------------------------------------------
 
-# 5.8.5.M1 (2022-07-11)
+# 5.8.5.M1 (2022-07-16)
 
 ### 🐣新特性
 * 【core   】     NumberUtil新增isIn方法（pr#669@Gitee）
@@ -14,6 +14,7 @@
 * 【core   】     新增CollectorUtil.reduceListMap()（pr#676@Gitee）
 * 【core   】     CollStreamUtil为空返回空的集合变为可编辑（pr#681@Gitee）
 * 【core   】     增加StrUtil.containsAll（pr#2437@Github）
+* 【core   】     ForestMap添加getNodeValue方法（pr#699@Gitee）
 * 
 ### 🐞Bug修复
 * 【core   】     修复CollUtil里面关于可变参数传null造成的crash问题（pr#2428@Github）